requirements.txt
pyjwt==2.1.0
fastapi[all]
pyyaml==5.4.1
tortoise-orm==0.17.4
passlib==1.7.4
bcrypt==3.2.0
asyncpg==0.23.0
psycopg2-binary==2.8.6
pandas==1.2.4